requirements.txt
# ============================================
# Django eCommerce Platform - Latest Packages
# Updated: 2024/2025
# ============================================

# Django Core (Latest LTS - 5.1)
Django==5.1.4
djangorestframework==3.15.2

# GraphQL Support
graphene-django==3.2.2
graphene==3.4.3
graphql-core==3.2.5
graphql-relay==3.2.0

# Database Drivers
mysqlclient==2.2.6
psycopg2-binary==2.9.10

# Django Extensions & Utilities
django-cors-headers==4.6.0
django-filter==24.3
django-environ==0.11.2

# Admin Interface
django-adminlte3==0.1.6
django-bootstrap4==24.4
django-crispy-forms==2.3
crispy-bootstrap4==2.0

# Content & Rich Text
django-ckeditor==6.7.1
django-js-asset==2.2.0

# Forms & Validation
django-countries==7.6.1
django-localflavor==4.0
django-simple-captcha==0.6.0
django-user-agents==0.4.0

# Email Integration
django-sendgrid-v5==1.2.3
sendgrid==6.11.0

# File & Image Handling
Pillow==11.0.0
django-ranged-response==0.2.0

# Payment Gateway Integrations
stripe==11.2.0
razorpay==1.4.2

# Security & Authentication
PyJWT==2.10.1
djangorestframework-simplejwt==5.3.1
django-allauth==65.3.0
twilio==9.3.7

# Date & Time Utilities
python-dateutil==2.9.0.post0
pytz==2024.2
python-stdnum==1.20

# HTTP & Web Scraping
requests==2.32.3
beautifulsoup4==4.12.3
soupsieve==2.6
lxml==5.3.0
selenium==4.27.1
webdriver-manager==4.0.2

# Static Files Management
whitenoise==6.8.2
brotli==1.1.0

# Currency Support
django-currencies==0.11.0
forex-python==1.8

# API Documentation
drf-spectacular==0.28.0
drf-spectacular-sidecar==2024.12.1

# ASGI & Performance
asgiref==3.8.1
uvicorn==0.34.0
gunicorn==23.0.0

# Async Task Queue (Celery)
celery==5.4.0
redis==5.2.1
django-celery-beat==2.7.0
django-celery-results==2.5.1

# Monitoring & Error Tracking
sentry-sdk==2.19.2

# Development & Testing
django-debug-toolbar==4.4.6
pytest-django==4.9.0
pytest==8.3.4
faker==33.1.0

# Additional Utilities
python-slugify==8.0.4
python-magic==0.4.27
urllib3==2.2.3
certifi==2024.12.14
charset-normalizer==3.4.0
idna==3.10
